+++ src/main/java/graveldb/datastore/lsmtree/memtable/MemtableStatus.java
@@ -1,6 +1,0 @@
-package graveldb.datastore.lsmtree.memtable;
-
-public enum MemtableStatus {
-    ACTIVE,
-    FLUSHED
-}