+++ src/main/java/graveldb/datastore/memtable/MemtableStatus.java
@@ -1,0 +1,8 @@
+package graveldb.datastore.memtable;
+
+public enum MemtableStatus {
+    ACTIVE,
+    TO_BE_FLUSHED,
+    FLUSHING,
+    FLUSHED
+}